cosmetic changes (line length mostly)

+++ srcs/utils/is_integral.hpp
@@ -6,7 +6,7 @@
 /*   By: twagner <twagner@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/06 09:52:59 by twagner           #+#    #+#             */
-/*   Updated: 2022/05/13 14:48:44 by twagner          ###   ########.fr       */
+/*   Updated: 2022/05/13 17:35:26 by twagner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,37 +28,41 @@ namespace ft
     typedef integral_constant<bool, false>  false_type;
 
     // Is_integral
-    template <class T>  struct is_integral                          : public false_type{};
-    template <>         struct is_integral<bool>                    : public true_type{};
-    template <>         struct is_integral<char>                    : public true_type{};
-    template <>         struct is_integral<char16_t>                : public true_type{};
-    template <>         struct is_integral<char32_t>                : public true_type{};
-    template <>         struct is_integral<wchar_t>                 : public true_type{};
-    template <>         struct is_integral<signed char>             : public true_type{};
-    template <>         struct is_integral<short int>               : public true_type{};
-    template <>         struct is_integral<int>                     : public true_type{};
-    template <>         struct is_integral<long int>                : public true_type{};
-    template <>         struct is_integral<long long int>           : public true_type{};
-    template <>         struct is_integral<unsigned char>           : public true_type{};
-    template <>         struct is_integral<unsigned short int>      : public true_type{};
-    template <>         struct is_integral<unsigned int>            : public true_type{};
-    template <>         struct is_integral<unsigned long int>       : public true_type{};
-    template <>         struct is_integral<unsigned long long int>  : public true_type{};
-    template <>         struct is_integral<const bool>                    : public true_type{};
-    template <>         struct is_integral<const char>                    : public true_type{};
-    template <>         struct is_integral<const char16_t>                : public true_type{};
-    template <>         struct is_integral<const char32_t>                : public true_type{};
-    template <>         struct is_integral<const wchar_t>                 : public true_type{};
-    template <>         struct is_integral<const signed char>             : public true_type{};
-    template <>         struct is_integral<const short int>               : public true_type{};
-    template <>         struct is_integral<const int>                     : public true_type{};
-    template <>         struct is_integral<const long int>                : public true_type{};
-    template <>         struct is_integral<const long long int>           : public true_type{};
-    template <>         struct is_integral<const unsigned char>           : public true_type{};
-    template <>         struct is_integral<const unsigned short int>      : public true_type{};
-    template <>         struct is_integral<const unsigned int>            : public true_type{};
-    template <>         struct is_integral<const unsigned long int>       : public true_type{};
-    template <>         struct is_integral<const unsigned long long int>  : public true_type{};
+    template <class T>
+    struct is_integral                                    : public false_type{};
+    template <> struct is_integral<bool>                   : public true_type{};
+    template <> struct is_integral<char>                   : public true_type{};
+    template <> struct is_integral<char16_t>               : public true_type{};
+    template <> struct is_integral<char32_t>               : public true_type{};
+    template <> struct is_integral<wchar_t>                : public true_type{};
+    template <> struct is_integral<signed char>            : public true_type{};
+    template <> struct is_integral<short int>              : public true_type{};
+    template <> struct is_integral<int>                    : public true_type{};
+    template <> struct is_integral<long int>               : public true_type{};
+    template <> struct is_integral<long long int>          : public true_type{};
+    template <> struct is_integral<unsigned char>          : public true_type{};
+    template <> struct is_integral<unsigned short int>     : public true_type{};
+    template <> struct is_integral<unsigned int>           : public true_type{};
+    template <> struct is_integral<unsigned long int>      : public true_type{};
+    template <> struct is_integral<unsigned long long int> : public true_type{};
+    template <> struct is_integral<const bool>             : public true_type{};
+    template <> struct is_integral<const char>             : public true_type{};
+    template <> struct is_integral<const char16_t>         : public true_type{};
+    template <> struct is_integral<const char32_t>         : public true_type{};
+    template <> struct is_integral<const wchar_t>          : public true_type{};
+    template <> struct is_integral<const signed char>      : public true_type{};
+    template <> struct is_integral<const short int>        : public true_type{};
+    template <> struct is_integral<const int>              : public true_type{};
+    template <> struct is_integral<const long int>         : public true_type{};
+    template <> struct is_integral<const long long int>    : public true_type{};
+    template <> struct is_integral<const unsigned char>    : public true_type{};
+    template <> struct is_integral<const unsigned int>     : public true_type{};
+    template <>
+    struct is_integral<const unsigned short int>           : public true_type{};
+    template <>
+    struct is_integral<const unsigned long int>            : public true_type{};
+    template <>
+    struct is_integral<const unsigned long long int>       : public true_type{};
 }
 
 #endif
